Fix dynamic_cast across dlopen boundaries.
Enabling _LIBCXX_DYNAMIC_FALLBACK allows type_info objects to be equal
if their strings are equal, not just their addresses. This is very
common on Android where libraries are loaded with dlopen via
System.loadLibrary.

This behavior matches libsupc++, which always does a string compare to
preserve compatibility for plugin architectures that behave similarly.

Gory details are in the bug.

Test: ndk/run_tests.py --filter dynamic_cast_dlopen
Bug: https://github.com/android-ndk/ndk/issues/533
Change-Id: I26cbf8d260cb7fb924580db6b346b42f39d5c2ed